"Unregister plugin translations when removed"
