+++ store/index.js
@@ -12,7 +12,7 @@ import configureServiceStore from 'mattermost-redux/store';
 import reduxInitialState from 'mattermost-redux/store/initial_state';
 
 import {storageRehydrate} from 'actions/storage';
-import {clearUserCookie} from 'actions/views/root';
+import {clearUserCookie} from 'actions/views/cookie';
 import appReducer from 'reducers';
 import {transformSet} from 'store/utils';
 import {detect} from 'utils/network.js';
